backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic>=2.8.0
pydantic-settings>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# Pin bcrypt to a compatible version for passlib to avoid __about__ import error
bcrypt==4.0.1
python-multipart==0.0.6
requests==2.31.0
pytesseract==0.3.10
pillow>=10.3.0
jinja2==3.1.2

gTTS==2.4.0

xgboost==2.0.3
scikit-learn>=1.3.2
pandas>=2.1.3
numpy>=1.26.2
python-dotenv==1.0.0
# Email services
email-validator
pyotp==2.9.0
psycopg2-binary==2.9.9
# LLM providers
google-generativeai==0.3.2

# PDF text extraction for OCR
pdfminer.six==20221105

# Real-time voice agent dependencies
vosk==0.3.44
supabase==2.3.4
joblib==1.3.2
websockets==12.0

# Piper TTS (local TTS)
piper-tts==1.3.0
groq
deepgram-sdk==5.3.0
faster-whisper
